Prevent app engine from charging me 5 quid a month
GAE works by scaling up instaces if they take more than {X}ms to
process a request. this is great and all for your Python applications
that suffer greatly under load and/or concurrency, but in our case
this actually means that any load will always case the blog to scale up.

So we are going to limit scaling to 1 instance, why is this okay?

A) We are behind aggressive cloudflare caching, the actual app
   does almost nothing, just making what lives in the CDN cache

B) Most of the latency we get from app engine is google itself!
   The GAE datastore  and cloud bucket actually take ~100ms to do
   things at times, very annoying! Since this almost defo means
   that we are going to scale up on ANY request.